Shutdown LLDB during a clean
This is a terrible hack.

We require these paths to be present for macos tests.

I believe that during the ‘build’ phase, lldb is already running (as evidenced by looking at the output of `ps -A | grep “lldb-rpc-server”`. Once lldb has been started, it’s already read in the contents of `~/.lldbinit`, and it’s too late to make changes.

Unfortunately, Xcode never shuts down the lldb instance. So, normally, this would mean users need to completely quit and re-open Xcode for changes to take place (meaning a full Xcode shutdown after every change to source - which is obviously a terrible situation - shutting down your IDE after every change - the velocity impact is huge).

Attempting to shut down lldb while building a project results in a run failure - because the running instance of lldb is the exact one Xcode will be using to start the process.

However, a shutdown during the clean phase works.

Also, a clean is automatic before a build with changes.