Stops trying to re-sign test bundles on macos.
Unit tests on macos do not require having been code signed. When an unsigned bundle is encountered (which is the default setup) an error occurs extracting the signing details.

An an alternative to this proposal, it may be better to check for an existing signature as a gating condition here instead.

This change is meant to address the problem with minimal broader-reaching side-effects.

ISSUES: https://github.com/bazelbuild/tulsi/issues/36